docbook-backend/requirements_BASE_1854.txt
django_heroku
django==3
requests
gunicorn
whitenoise
django-rest-framework
django_extensions
